Allow enabling WDS on the private WiFi feature.
When distributing a wireless local network in addition to Freifunk, the
existing wireless uplink functionality can be used to connect to a base AP
and bridge the WAN interface of the router to a wireless uplink. This
adaptation in the Private WiFi feature allows creating an access point
usable for WDS frames (4-address-frames), so that the wireless uplink can
directly be connected to another Freifunk router without manually changing
the configuration.